backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.1
sqlalchemy[asyncio]>=2.0.38
alembic>=1.15.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
httpx>=0.28.1
slowapi>=0.1.9
pydantic-settings>=2.7.1
python-multipart>=0.0.20
python-dotenv>=1.0.1
aiosqlite>=0.20.0
python-dateutil>=2.9.0
